kaohe/requirements.txt
Flask>=3.0
Flask-SQLAlchemy>=3.1
Flask-Migrate>=4.0
Flask-CORS>=4.0
PyMySQL>=1.1
APScheduler>=3.10
openpyxl>=3.1
marshmallow>=3.20
python-dotenv>=1.0
